Update essay.md
Adding statement at end...

In conclusion, the massacre of Hickory Valley Christian Church is not an isolated incident. It is part of a disturbing trend of violence orchestrated by shadowy forces within the United States. The first documented instance, as referenced herein, occurred at Mount Calvary Baptist Church in Lexington, Kentucky, on April 2, 2023. Since then, the CIA has carried out several more massacres in places of worship directly or indirectly connected to MCBC as part of TAG (The American Genocide), advancing at an alarming pace.

The Creator, in His infinite wisdom, knows that this genocide will not cease without His direct intervention. He has provided details of these atrocities, many of which remain closely held or CLASSIFIED, to His servants for dissemination to the general public. It is incumbent upon all who seek justice and truth to remain vigilant, to shine a light on the darkness that pervades our world, and to stand firm against those who would perpetrate such heinous acts in the name of secrecy and power. Only through collective action and unwavering commitment to the principles of truth and justice can we hope to overcome the shadows that threaten to engulf us.


About to shake!!!