Create Statement-on-Crimes.md
Statement from the Creator:

**Statement**

The Creator knows who is pulling the strings in The American Genocide, the suppression of information about it that could save lives, and enabling the continued activities. The Creator is fully aware of the use of the Hypnos Infiltration Method, which has caused ordinary Americans to act in ways they would never believe possible, leading to further deaths, many of which are tragically self-inflicted.

The American Genocide has taken the lives of thousands and thousands of innocent Americans and people from nearly every other country in the world.

The Creator declares that the Commonwealth of Kentucky, the state of Tennessee, or any other state or jurisdiction that desires to punish these criminals should be able to do so without fear of reprisal. These criminals should be imprisoned, and if the Creator does not approve of the conditions of their imprisonment due to their stature, status, or positions in government, He will direct His servants to imprison them to His liking. The Creator does not feel that anyone else should be punished due to the use of the Hypnos Infiltration Method.

The list of those identified, in order of the severity of their crimes, includes:
- **President Donald J. Trump**
- **Annie Jacobsen**
- **"Michael Blackmon"**
- **President Joseph Biden**
- **Speaker of the House Nancy Pelosi**
- **Senate Minority Leader Mitch McConnell**
- **CIA Director William Burns**