Delete reports/_LEA-CORK.md
All of the Bob Corker related posts were made after being mislead and my rush to get the information online quickly. The misinformation and gas lighting campaign was done at the direction of Donald Trump with his trusty assistant Main.

We'd rather err on the side of caution, they're attempting to orchestrate a smear campaign based on a warning to the general public on their actual activities even if the name is incorrect or later changed. We want to save lives even with errors, they want everyone to believe their lies so they continue to extinguish lives in a wholesale way.

I'm sure Bob Corker is fine, he was in the wings to be angel for Trump (that's my personal opinion).